split up server and add index.js
