+++ server/server.js
@@ -1,16 +1,18 @@
-// Data feed imports
-const { sectorInformation } = require("./components/sectorInformation");
-const { topPeers } = require("./components/topPeers");
-const { companyOverview } = require("./components/companyOverview");
-const { latestNewsInterval } = require("./components/latestNewsInterval");
-const { keyStats } = require("./components/keyStats");
-const { stockTicker } = require("./components/stockTicker");
-const { chartData } = require("./components/chartData");
+const {
+  sectorInformation,
+  topPeers,
+  companyOverview,
+  latestNewsInterval,
+  keyStats,
+  stockTicker,
+  chartData,
+  searchQuery,
+  getAllCompanys
+} = require("./components");
 
 const express = require("express");
 const http = require("http");
 const socketIo = require("socket.io");
-const axios = require("axios");
 const port = process.env.PORT || 4000;
 const index = require("./routes");
 const app = express();
@@ -18,6 +20,8 @@ app.use(index);
 const server = http.createServer(app);
 const io = socketIo(server);
 
+const HOST = "https://sandbox.iexapis.com/stable";
+const TOKEN = "Tsk_d2f1890612194476b41d39992a3ad835";
 const oneDay = 24 * 60 * 60 * 1000;
 
 function callAndStartIntervals(fn, interval, ...args) {
@@ -27,7 +31,7 @@ function callAndStartIntervals(fn, interval, ...args) {
 
 io.on("connection", socket => {
   const timerIDs = {};
-  const allSymbols = getCompanySymbols();
+  const allSymbols = getAllCompanys(HOST, TOKEN);
   console.info("New client connected");
   socket.on("stockName", async (stockName, timeRange) => {
     if (stockName === "") {
@@ -112,37 +116,3 @@ io.on("connection", socket => {
 });
 
 server.listen(port, () => console.log(`Listening on port ${port}`));
-
-const HOST = "https://sandbox.iexapis.com/stable";
-const TOKEN = "Tsk_d2f1890612194476b41d39992a3ad835";
-
-const getCompanySymbols = async () => {
-  try {
-    const companySymbols = await axios.get(
-      `${HOST}/ref-data/symbols?token=${TOKEN}`
-    );
-
-    return companySymbols.data.map(data => ({
-      symbol: data.symbol,
-      name: data.name,
-      exchange: data.exchange
-    }));
-  } catch (error) {
-    console.error(`Error: ${error}`);
-  }
-};
-
-const searchQuery = async (socket, inputQuery, allSymbols) => {
-  try {
-    const symbols = await allSymbols;
-    const filteredData = symbols.filter(
-      search =>
-        search.symbol.toLowerCase().indexOf(inputQuery.toLowerCase()) !== -1 ||
-        search.name.toLowerCase().indexOf(inputQuery.toLowerCase()) !== -1
-    );
-    const topTen = filteredData.slice(0, 10);
-    socket.emit("companySymbols", topTen);
-  } catch (error) {
-    console.error(`Error: ${error}`);
-  }
-};
